Moving the etcd wrapper into the etcd folder
